backend/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
pytest
pytest-django
pytest-cov
pytest-xdist
flake8
black
isort
django-debug-toolbar
